scrapper/requirements.txt
scrapy==2.12.0
playwright==1.49.1
scrapy-playwright==0.0.43
beautifulsoup4==4.12.3
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
celery[redis]==5.4.0
pyyaml==6.0.2
redis==5.2.1
